Refactoring service
